feat: 게임방 생성 구현

+++ src/main/java/com/ureca/miniproject/user/repository/UserRepository.java
@@ -1,12 +1,9 @@
 package com.ureca.miniproject.user.repository;
 
-import java.util.Optional;
-
+import com.ureca.miniproject.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.ureca.miniproject.user.entity.User;
-
-public interface UserRepository extends JpaRepository<User, Integer>{
+public interface UserRepository extends JpaRepository<User, Long> {
 	//crud 생성 완료
 //	Optional<User> findByEmail(String email);
 }
